Fix issues with hypervisor detection on hypervisors that print more than one line
Replace `head` and `tail` for disecting output of `virt-what` with `awk`. While this seems to be a lot of overhead we can not be sure about different outputs on different hypervisors. We need more tests before redoing the detection as a whole.

This helps with #1 and #15 but is still not a final state