Add check for non-endpoint zones being synced
See https://github.com/Icinga/icinga2/issues/7530 for the background of this check

It checks if zones that are not mere check_endpoint agents but receive configuration aside from global configuration (e.g. satellites, agents with local scheduler) have their Zone object lying under `/etc/icinga2/zones.d` instead of in `/etc/icinga2/zones.conf` where it should belong.

Thanks to https://stackoverflow.com/questions/16854280/a-variable-modified-inside-a-while-loop-is-not-remembered for the nice solution with the here-document

fixes #87